Fixed that the cats move out of the boundaries.
The program now has no problem to move the dog in response to user's
control. The cats also reacts to the dog's position automatically
according to their rules(1-6) of movement. Programs achieves 95%
completeness.

- Fixed that Cats are not within the boundaries already when they are
generated in the first place.
- Updated Cat class's methods that are relevant to 'move' functions,
e.g., moveUp method, moveToAdajacent method and so on.
- Fixed the issue in inner class WindowResizingHanlder in
ViewModelController class. The program now only reacts to the up, down,
left, right keys in keyboard.