Change the rule of cat's movement and fixed the issue of restarting game
From the game design perspective, the rule 6 (which asks the cat to move
to the square of the dog) just doesn't make sense at all. I choose to
comment it.

- removing rule 6 in move method
- commented moveToDog method

When the user wins the game and selects yes to restart the game, the new
cats and dog are created, and their positions on the GUI are also
updated.